requirements.txt
python-pptx
google-generativeai
python-dotenv 